requirements.txt
astroid==2.2.5
autopep8==1.4.4
Django==2.2.3
django-cors-headers==3.0.2
django-extensions==2.1.9
django-rest-auth==0.9.5
djangorestframework==3.9.4
djangorestframework-jwt==1.11.0
isort==4.3.21
lazy-object-proxy==1.4.1
mccabe==0.6.1
Pillow==6.1.0
psycopg2==2.8.3
pycodestyle==2.5.0
PyJWT==1.7.1
pylint==2.3.1
pylint-django==2.0.11
pylint-plugin-utils==0.5
pytz==2019.1
six==1.12.0
sqlparse==0.3.0
typed-ast==1.4.0
wrapt==1.11.2
